+++ src/rules/layout-token-use/index.js
@@ -52,7 +52,7 @@ const defaultOptions = {
   acceptFluidSpacingTokens: false,
   acceptCarbonMiniUnitsFunction: false,
   acceptScopes: ["layout"],
-  testOnlyTarget: undefined
+  testOnlyVersion: undefined
 };
 
 export default function rule(primaryOptions, secondaryOptions, context) {
@@ -81,8 +81,8 @@ export default function rule(primaryOptions, secondaryOptions, context) {
             val === undefined || typeof val === "boolean",
           acceptCarbonMiniUnitsFunction: (val) =>
             val === undefined || typeof val === "boolean",
-          testOnlyTarget: (val) =>
-            val === undefined || ["v10", "v11"].includes(val)
+          testOnlyVersion: (val) =>
+            val === undefined || ["10", "v11"].includes(val)
         },
         optional: true
       }