+++ src/rules/motion-token-use/utils/fixes.js
@@ -10,25 +10,25 @@ import { fixUsingMap } from "../../../utils/fixUsingMap";
 export const fixes = [
   {
     // remove carbon prefix from spacing tokens
-    version: "v11",
+    version: "11",
     target: /\$(duration--)?((fast)|(moderate)|(slow))-([0-2]{2})/g,
     replacement: "$$duration-$2-$6"
   },
   {
     // remove carbon prefix for easings
-    version: "v11",
+    version: "11",
     target: /\$(carbon--)((ease-in)|(ease-out)|(standard-easing)|(easings))/g,
     replacement: "$$$2"
   },
   {
     // remove carbon prefix for motion function
-    version: "v11",
+    version: "11",
     target: /carbon--motion/g,
     replacement: "motion"
   },
   {
     // replace matching motion duration literal
-    version: "v11",
+    version: "11",
     target: /[0-9]*ms/g,
     replacement: (value, target) => {
       return fixUsingMap(value, target, {
@@ -38,6 +38,21 @@ export const fixes = [
         "240ms": "$duration-moderate-02",
         "400ms": "$duration-slow-01",
         "700ms": "$duration-slow-02"
+      });
+    }
+  },
+  {
+    // replace matching motion duration literal
+    version: "10",
+    target: /[0-9]*ms/g,
+    replacement: (value, target) => {
+      return fixUsingMap(value, target, {
+        "70ms": "$duration--fast-01",
+        "110ms": "$duration--fast-02",
+        "150ms": "$duration--moderate-01",
+        "240ms": "$duration--moderate-02",
+        "400ms": "$duration--slow-01",
+        "700ms": "$duration--slow-02"
       });
     }
   }