+++ src/rules/motion-token-use/utils/getMotionInfo.js
@@ -9,8 +9,8 @@ import { doInit } from "./initMotionTokens";
 import { fixes } from "./fixes";
 
 export default async function getMotionInfo(options) {
-  const { motionTokens, motionFunctions } = await doInit(
-    options.testOnlyTarget
+  const { motionTokens, motionFunctions, version } = await doInit(
+    options.testOnlyVersion
   );
 
   return {
@@ -28,6 +28,7 @@ export default async function getMotionInfo(options) {
         values: motionFunctions
       }
     ],
-    fixes
+    fixes,
+    version
   };
 }